+++ src/snowflake/snowpark/modin/plugin/_internal/binary_op_utils.py
@@ -4,14 +4,29 @@
 import functools
 from collections.abc import Hashable
 from dataclasses import dataclass
+from enum import Enum, auto
 
+import pandas as native_pd
 from pandas._typing import Callable, Scalar
 
 from snowflake.snowpark.column import Column as SnowparkColumn
-from snowflake.snowpark.functions import col, concat, floor, iff, repeat, when
+from snowflake.snowpark.functions import (
+    col,
+    concat,
+    datediff,
+    floor,
+    iff,
+    is_null,
+    repeat,
+    when,
+)
 from snowflake.snowpark.modin.plugin._internal.frame import InternalFrame
 from snowflake.snowpark.modin.plugin._internal.join_utils import (
     JoinOrAlignInternalFrameResult,
+)
+from snowflake.snowpark.modin.plugin._internal.snowpark_pandas_types import (
+    SnowparkPandasColumn,
+    TimedeltaType,
 )
 from snowflake.snowpark.modin.plugin._internal.type_utils import infer_object_type
 from snowflake.snowpark.modin.plugin._internal.utils import pandas_lit
@@ -20,6 +35,8 @@ from snowflake.snowpark.types import (
     DataType,
     NullType,
     StringType,
+    TimestampTimeZone,
+    TimestampType,
     _FractionalType,
     _IntegralType,
 )
@@ -172,13 +189,64 @@ def is_binary_op_supported(op: str) -> bool:
     return op in SUPPORTED_BINARY_OPERATIONS
 
 
+class SubtractionType(Enum):
+    """Type of subtraction, i.e. rsub or sub"""
+
+    # SUB means first_operand - second_operand, e.g. pd.Series(2).sub(pd.Series(1)) is equal to pd.Series(1)
+    SUB = auto()
+    # RSUB means second_operand - first_operand, e.g. pd.Series(2).rsub(pd.Series(1)) is equal to pd.Series(-1)
+    RSUB = auto()
+
+
+def _compute_subtraction_between_snowpark_timestamp_columns(
+    first_operand: SnowparkColumn,
+    first_datatype: DataType,
+    second_operand: SnowparkColumn,
+    second_datatype: DataType,
+    subtraction_type: SubtractionType,
+) -> SnowparkPandasColumn:
+    """
+    Compute subtraction between two snowpark columns.
+
+    Args:
+        first_operand: SnowparkColumn for lhs
+        first_datatype: Snowpark datatype for lhs
+        second_operand: SnowparkColumn for rhs
+        second_datatype: Snowpark datatype for rhs
+        subtraction_type: Type of subtraction.
+    """
+    if (
+        first_datatype.tz is TimestampTimeZone.NTZ
+        and second_datatype.tz is TimestampTimeZone.TZ
+    ) or (
+        first_datatype.tz is TimestampTimeZone.TZ
+        and second_datatype.tz is TimestampTimeZone.NTZ
+    ):
+        raise TypeError("Cannot subtract tz-naive and tz-aware datetime-like objects.")
+    return SnowparkPandasColumn(
+        iff(
+            is_null(first_operand).__or__(is_null(second_operand)),
+            pandas_lit(native_pd.NaT),
+            datediff(
+                "ns",
+                *(
+                    (first_operand, second_operand)
+                    if subtraction_type is SubtractionType.RSUB
+                    else (second_operand, first_operand)
+                ),
+            ),
+        ),
+        TimedeltaType(),
+    )
+
+
 def compute_binary_op_between_snowpark_columns(
     op: str,
     first_operand: SnowparkColumn,
     first_datatype: Callable[[], DataType],
     second_operand: SnowparkColumn,
     second_datatype: Callable[[], DataType],
-) -> SnowparkColumn:
+) -> SnowparkPandasColumn:
     """
     Compute pandas binary operation for two SnowparkColumns
     Args:
@@ -191,10 +259,9 @@ def compute_binary_op_between_snowpark_columns(
         it is not needed.
 
     Returns:
-        SnowparkColumn expr for translated pandas operation
+        SnowparkPandasColumn for translated pandas operation
     """
-
-    binary_op_result_column = None
+    binary_op_result_column, snowpark_pandas_result_type = None, None
 
     # some operators and the data types have to be handled specially to align with pandas
     # However, it is difficult to fail early if the arithmetic operator is not compatible
@@ -272,14 +339,48 @@ def compute_binary_op_between_snowpark_columns(
             binary_op_result_column = pandas_lit(False)
         else:
             binary_op_result_column = first_operand.equal_null(second_operand)
-
+    elif (
+        op in ("rsub", "sub")
+        and isinstance(first_datatype(), TimestampType)
+        and isinstance(second_datatype(), NullType)
+    ):
+        # Timestamp - NULL or NULL - Timestamp raises SQL compilation error,
+        # but it's valid in pandas and returns NULL.
+        snowpark_pandas_result_type = NullType()
+        binary_op_result_column = pandas_lit(None)
+    elif (
+        op in ("rsub", "sub")
+        and isinstance(first_datatype(), NullType)
+        and isinstance(second_datatype(), TimestampType)
+    ):
+        # Timestamp - NULL or NULL - Timestamp raises SQL compilation error,
+        # but it's valid in pandas and returns NULL.
+        snowpark_pandas_result_type = NullType()
+        binary_op_result_column = pandas_lit(None)
+    elif (
+        op in ("sub", "rsub")
+        and isinstance(first_datatype(), TimestampType)
+        and isinstance(second_datatype(), TimestampType)
+    ):
+        return _compute_subtraction_between_snowpark_timestamp_columns(
+            first_operand=first_operand,
+            first_datatype=first_datatype(),
+            second_operand=second_operand,
+            second_datatype=second_datatype(),
+            subtraction_type=SubtractionType.SUB
+            if op == "sub"
+            else SubtractionType.RSUB,
+        )
     # If there is no special binary_op_result_column result, it means the operator and
     # the data type of the column don't need special handling. Then we get the overloaded
     # operator from Snowpark Column class, e.g., __add__ to perform binary operations.
     if binary_op_result_column is None:
         binary_op_result_column = getattr(first_operand, f"__{op}__")(second_operand)
 
-    return binary_op_result_column
+    return SnowparkPandasColumn(
+        snowpark_column=binary_op_result_column,
+        snowpark_pandas_type=snowpark_pandas_result_type,
+    )
 
 
 def are_equal_types(type1: DataType, type2: DataType) -> bool:
@@ -307,7 +408,7 @@ def compute_binary_op_between_snowpark_column_and_scalar(
     first_operand: SnowparkColumn,
     datatype: Callable[[], DataType],
     second_operand: Scalar,
-) -> SnowparkColumn:
+) -> SnowparkPandasColumn:
     """
     Compute the binary operation between a Snowpark column and a scalar.
     Args:
@@ -318,16 +419,14 @@ def compute_binary_op_between_snowpark_column_and_scalar(
         second_operand: Scalar value
 
     Returns:
-        The result as a Snowpark column
+        SnowparkPandasColumn for translated pandas operation
     """
 
     def second_datatype() -> DataType:
         return infer_object_type(second_operand)
 
-    second_operand = pandas_lit(second_operand)
-
     return compute_binary_op_between_snowpark_columns(
-        op, first_operand, datatype, second_operand, second_datatype
+        op, first_operand, datatype, pandas_lit(second_operand), second_datatype
     )
 
 
@@ -336,7 +435,7 @@ def compute_binary_op_between_scalar_and_snowpark_column(
     first_operand: Scalar,
     second_operand: SnowparkColumn,
     datatype: Callable[[], DataType],
-) -> SnowparkColumn:
+) -> SnowparkPandasColumn:
     """
     Compute the binary operation between a scalar and a Snowpark column.
     Args:
@@ -347,16 +446,14 @@ def compute_binary_op_between_scalar_and_snowpark_column(
         it is not needed.
 
     Returns:
-        The result as a Snowpark column
+        SnowparkPandasColumn for translated pandas operation
     """
 
     def first_datatype() -> DataType:
         return infer_object_type(first_operand)
 
-    first_operand = pandas_lit(first_operand)
-
     return compute_binary_op_between_snowpark_columns(
-        op, first_operand, first_datatype, second_operand, datatype
+        op, pandas_lit(first_operand), first_datatype, second_operand, datatype
     )
 
 
@@ -367,7 +464,7 @@ def compute_binary_op_with_fill_value(
     rhs: SnowparkColumn,
     rhs_datatype: Callable[[], DataType],
     fill_value: Scalar,
-) -> SnowparkColumn:
+) -> SnowparkPandasColumn:
     """
     Helper method for performing binary operations.
     1. Fills NaN/None values in the lhs and rhs with the given fill_value.
@@ -392,7 +489,7 @@ def compute_binary_op_with_fill_value(
             successful DataFrame alignment, with this value before computation.
 
     Returns:
-        SnowparkColumn expression for translated pandas operation
+        SnowparkPandasColumn for translated pandas operation
     """
     lhs_cond, rhs_cond = lhs, rhs
     if fill_value is not None: